djoserAuth/django_polls_auth/requirements.txt
Django==1.8
djangorestframework==3.3.3
